riscv: cheri: read capability from kernel stack
Update regs_get_kernel_stack_nth to return a capability.
This function is called from kprobes if a probe wants to read a
stack word.

+++ arch/riscv/include/asm/ptrace.h
@@ -133,7 +133,7 @@ static inline void regs_set_return_value(struct pt_regs *regs,
 }
 
 extern int regs_query_register_offset(const char *name);
-extern unsigned long regs_get_kernel_stack_nth(struct pt_regs *regs,
+extern uintptr_t regs_get_kernel_stack_nth(struct pt_regs *regs,
 					       unsigned int n);
 
 void prepare_ftrace_return(uintptr_t *parent, unsigned long self_addr,
